migrate gravity effect

+++ lib/Commands/Commands_kaleidoscope.cpp
@@ -4,7 +4,8 @@ void Commands::kaleidoscope(LedStructure *leds, char *data)
 {
   uint8_t strip_index = data[0];
 
-  uint8_t length = leds->strip_length(strip_index);
+  uint16_t length = leds->strip_length(strip_index);
+  
   for (int i = 0; i < length / 2; i++)
   {
     leds->set_led(strip_index, length - i - 1, leds->get_led(strip_index, i));
